feat: refactor + new folders + clean up
