+++ vue/src/components/sensor/SensorComp.vue
