+++ vue/src/views/CropImageView.vue
@@ -1,10 +1,10 @@
 <template>
   <base-layout title="Crop Image">
-    <!-- <image-crop></image-crop> -->
-    <image-crop-quasar></image-crop-quasar>
+    <image-crop></image-crop>
   </base-layout>
 </template>
 
 <script setup lang="ts">
 import BaseLayout from "@/components/layout/BaseLayout.vue";
+import ImageCrop from "@/components/image-crop/ImageCrop.vue";
 </script>
